# api/docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/rides": {
      "post": {
        "operationId": "RidesController_create",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRideDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or scheduling rule violation."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Create a ride template in the current tenant.",
        "tags": [
          "Rides"
        ]
      },
      "get": {
        "operationId": "RidesController_list",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "default": 25,
              "example": 25,
              "type": "number"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "RECURRING",
                "ONE_TIME"
              ]
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "DRAFT",
                "ACTIVE",
                "INACTIVE"
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "central",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedRidesResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List rides in the current tenant.",
        "tags": [
          "Rides"
        ]
      }
    },
    "/rides/instances": {
      "get": {
        "operationId": "RidesController_listInstancesByDate",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "date",
            "required": true,
            "in": "query",
            "schema": {
              "example": "2026-03-30",
              "type": "string"
            }
          },
          {
            "name": "timezoneOffsetMinutes",
            "required": false,
            "in": "query",
            "description": "Timezone offset in minutes, where west of UTC is negative and east of UTC is positive.",
            "schema": {
              "minimum": -840,
              "maximum": 840,
              "example": 0,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideInstancesByDateResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "List materialized ride instances by date in the current tenant.",
        "tags": [
          "Rides"
        ]
      }
    },
    "/rides/{id}": {
      "get": {
        "operationId": "RidesController_getById",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Ride not found in current tenant."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Get a ride template detail by id in the current tenant.",
        "tags": [
          "Rides"
        ]
      },
      "patch": {
        "operationId": "RidesController_update",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRideDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or scheduling rule violation."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Ride not found in current tenant."
          },
          "409": {
            "description": "The change would break existing reservations, such as lowering capacity under a seat already sold. Resend with confirmBreakingChange to proceed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Update ride fields in the current tenant.",
        "tags": [
          "Rides"
        ]
      },
      "put": {
        "operationId": "RidesController_replace",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRideDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or scheduling rule violation."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Ride not found in current tenant."
          },
          "409": {
            "description": "The change would break existing reservations, such as lowering capacity under a seat already sold. Resend with confirmBreakingChange to proceed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Replace ride fields in the current tenant.",
        "tags": [
          "Rides"
        ]
      },
      "delete": {
        "operationId": "RidesController_remove",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cascade",
            "required": false,
            "in": "query",
            "description": "When true, cancels active reservations and deactivates the ride.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Ride not found in current tenant."
          },
          "409": {
            "description": "Ride has active reservations and cascade override is not enabled."
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a ride template in the current tenant.",
        "tags": [
          "Rides"
        ]
      }
    },
    "/rides/{id}/day-times": {
      "put": {
        "operationId": "RidesController_replaceDayTimes",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplaceRideDaySchedulesDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or scheduling rule violation."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Ride not found in current tenant."
          },
          "409": {
            "description": "The proposed schedule would break an existing reservation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Replace recurring ride day schedules in the current tenant.",
        "tags": [
          "Rides"
        ]
      }
    },
    "/rides/{id}/exceptions": {
      "post": {
        "operationId": "RidesController_addException",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRideExceptionDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideExceptionResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failure or invalid exception combination."
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Ride not found in current tenant."
          },
          "409": {
            "description": "Either a SKIP that would break a reservation, carrying WOULD_BREAK_RESERVATIONS and confirmable, or a duplicate or conflicting exception, which is not.",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/WouldBreakReservationsDto"
                    },
                    {
                      "$ref": "#/components/schemas/RideExceptionConflictDto"
                    }
                  ]
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Add a ride exception (skip/additional) in the current tenant.",
        "tags": [
          "Rides"
        ]
      }
    },
    "/rides/{id}/exceptions/{exceptionId}": {
      "delete": {
        "operationId": "RidesController_removeException",
        "parameters": [
          {
            "name": "X-Tenant-Slug",
            "in": "header",
            "description": "Tenant slug that must match authenticated token tenant.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exceptionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmBreakingChangeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RideExceptionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token."
          },
          "403": {
            "description": "Insufficient role for this resource."
          },
          "404": {
            "description": "Ride or exception not found in current tenant."
          },
          "409": {
            "description": "Deleting the exception would break an existing reservation.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WouldBreakReservationsDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "summary": "Delete a ride exception in the current tenant.",
        "tags": [
          "Rides"
        ]
      }
    },
...
  },
  "info": {
    "title": "SURP API",
    "description": "Backend API for SURP, providing tenant-scoped authentication, operational health endpoints, and secure session management for transportation workflows. Protected endpoints require BOTH Authorization: Bearer <accessToken> and X-Tenant-Slug headers (except /platform/* routes), and the tenant must match the token claim.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token issued by /auth/login. Use together with X-Tenant-Slug on protected endpoints."
      }
    },
    "schemas": {
...
      "CreateRideDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "Morning Central Route"
          },
          "lineId": {
            "type": "string",
            "example": "line-id-123"
          },
          "capacity": {
            "type": "number",
            "example": 38,
            "minimum": 1,
            "maximum": 100
          },
          "type": {
            "type": "string",
            "enum": [
              "RECURRING",
              "ONE_TIME"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "INACTIVE"
            ],
            "default": "DRAFT"
          },
          "recurringStartDate": {
            "type": "string",
            "example": "2026-03-14"
          },
          "recurringEndDate": {
            "type": "string",
            "example": "2026-06-30"
          },
          "oneTimeDate": {
            "type": "string",
            "example": "2026-03-20"
          },
          "oneTimeDepartureTime": {
            "type": "string",
            "example": "09:00"
          },
          "oneTimeArrivalTime": {
            "type": "string",
            "example": "10:30"
          },
          "daySchedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RideDayScheduleInputDto"
            }
          }
        },
        "required": [
          "lineId",
          "capacity",
          "type"
        ]
      },
...
      "RideResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "tenantId": {
            "type": "string"
          },
          "lineId": {
            "type": "string"
          },
          "createdById": {
            "type": "object",
            "nullable": true
          },
          "updatedById": {
            "type": "object",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "capacity": {
            "type": "number",
            "minimum": 1
          },
          "type": {
            "type": "string",
            "enum": [
              "RECURRING",
              "ONE_TIME"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "DRAFT",
              "ACTIVE",
              "INACTIVE"
            ]
          },
          "recurringStartDate": {
            "type": "object",
            "nullable": true,
            "example": "2026-03-14"
          },
          "recurringEndDate": {
            "type": "object",
            "nullable": true,
            "example": "2026-06-30"
          },
          "oneTimeDate": {
            "type": "object",
            "nullable": true,
            "example": "2026-03-20"
          },
          "oneTimeDepartureTime": {
            "type": "object",
            "nullable": true
          },
          "oneTimeArrivalTime": {
            "type": "object",
            "nullable": true
          },
          "daySchedules": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RideDayScheduleResponseDto"
            }
          },
          "exceptions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RideExceptionResponseDto"
            }
          },
          "line": {
            "$ref": "#/components/schemas/RideLineSummaryDto"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "tenantId",
          "lineId",
          "createdById",
          "updatedById",
          "name",
          "capacity",
          "type",
          "status",
          "recurringStartDate",
          "recurringEndDate",
          "oneTimeDate",
          "oneTimeDepartureTime",
          "oneTimeArrivalTime",
          "daySchedules",
          "exceptions",
          "line",
          "createdAt",
          "updatedAt"
        ]
      },
      "PaginatedRidesResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RideResponseDto"
            }
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "pageSize": {
            "type": "number",
            "example": 25
          },
          "total": {
            "type": "number",
            "example": 2
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total"
        ]
      },
...
    }
  }
}